aws:PrincipalTag gets tags from the IAM role, not from the EC2 instance
I can't use EC2 instance tags to control access to things :/. That's
lame. I think in production it'll be okay, since really all I'm doing
is associating per service roles to instances. Here I'm just using
the AMI name as the service name to be "clever" and get a 1:1 mapping
from AMI to logs from the server that made it.